backend/requirements.txt
# requirements.txt
fastapi==0.95.0
fastapi-login
uvicorn==0.23.1
llama-index==0.5.6
openai==0.27.0
unstructured==0.6.5
faiss-cpu==1.7.4
redis==4.5.5
psycopg2-binary==2.9.6
elasticsearch==8.10.0
boto3==1.26.7
pydantic
python-dotenv==0.21.1
jinja2==3.1.2
celery==5.2.6
gunicorn==20.1.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
databases
asyncpg
autogen
pdf2image==1.16.3
pytesseract==0.3.10
nltk==3.6.7
pandas==1.3.3
requests==2.26.0
tqdm
pymupdf
beautifulsoup4
scikit-learn
google-generativeai==0.8.3
alembic==1.7.7
langchain
langchain-community
spacy